Make API responses REST-like

+++ src/main/java/fi/oispakaljaa/karhu/controller/BarController.java
@@ -1,9 +1,12 @@
 package fi.oispakaljaa.karhu.controller;
 
+import fi.oispakaljaa.karhu.APItemplates.google.OispakaljaaTemplate;
 import fi.oispakaljaa.karhu.domain.Bar;
 import fi.oispakaljaa.karhu.repository.BarRepository;
-import java.util.Collection;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -11,33 +14,40 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping(value = "bars")
+@RequestMapping(value = "/api/bars")
 public class BarController {
 
     @Autowired
     BarRepository barRepository;
 
     @RequestMapping(method = RequestMethod.GET)
-    public Collection<Bar> list() {
-        return barRepository.findAll();
+    public ResponseEntity<OispakaljaaTemplate> list() {
+        return new ResponseEntity<>(new OispakaljaaTemplate("OK", "", barRepository.findAll()), HttpStatus.OK);
     }
 
     @RequestMapping(value = "/{barId}", method = RequestMethod.GET)
-    public Bar getBar(@PathVariable("barId") Long id) {
-        return barRepository.findOne(id);
+    public ResponseEntity<OispakaljaaTemplate> getBar(@PathVariable("barId") Long id) {
+        Bar b = barRepository.findOne(id);
+        if (b == null)
+            return new ResponseEntity<>(new OispakaljaaTemplate("Error", "Bar not found", null), HttpStatus.NOT_FOUND);
+
+        return new ResponseEntity<>(new OispakaljaaTemplate("OK", "", b), HttpStatus.OK);
     }
 
     @RequestMapping(method = RequestMethod.POST)
-    public Bar postBar(@RequestBody Bar bar) {
-        return barRepository.save(bar);
+    public ResponseEntity<OispakaljaaTemplate> postBar(@RequestBody Bar bar) {
+        return new ResponseEntity<OispakaljaaTemplate>(new OispakaljaaTemplate("OK", "", barRepository.save(bar)), HttpStatus.OK);
     }
 
     @RequestMapping(value = "/{barId}", method = RequestMethod.DELETE)
-    public Bar deleteBar(@PathVariable("barId") Long id) {
+    public ResponseEntity<OispakaljaaTemplate> deleteBar(@PathVariable("barId") Long id) {
         Bar bar = barRepository.findOne(id);
+        if (bar == null)
+            return new ResponseEntity<>(new OispakaljaaTemplate("Error", "Bar not found", null), HttpStatus.NOT_FOUND);
+
         barRepository.delete(bar);
-        return bar;
-        
+        return new ResponseEntity<OispakaljaaTemplate>(new OispakaljaaTemplate("OK", "", bar), HttpStatus.OK);
+
         // se täytyis varmistaa, että jos baari poistetaan niin myös kaikki kyseisen baarin juomat poistetaan
     }
